+++ Códigos/DiagnosticoSaude/src/Dao/PacienteDAO.java
@@ -1,72 +1,72 @@
-package Model;
+package Dao;
 
-import java.sql.Connection;
-import java.sql.SQLException;
+import Model.Paciente;
 import java.sql.PreparedStatement;
+import java.sql.SQLException;
 import java.util.Calendar;
 import java.sql.Date;
 import java.sql.ResultSet;
 import java.util.ArrayList;
 import java.util.List;
 
-public class EnfermeiroDAO<Object> extends DAO<Enfermeiro> {
+public class PacienteDAO<Object> extends DAO<Paciente> {
 
-    public EnfermeiroDAO() throws SQLException {
+    public PacienteDAO() throws SQLException {
         super();
     }
 
     @Override
-    public boolean add(Enfermeiro modelo) throws SQLException {
+    public boolean add(Paciente modelo) throws SQLException {
         if (searchBy("id", modelo.getId()).size() > 0) {
             return false;
         }
 
-        PreparedStatement stmt = con.prepareStatement("insert into enfermeiro (id, nome, nascimento, sexo, login, senha) values (?, ?, ?, ?, ?, ?)");
+        PreparedStatement stmt = con.prepareStatement("insert into paciente (id, nome, nascimento, sexo, enfermeiro) values (?, ?, ?, ?, ?)");
         stmt.setString(1, modelo.getId());
         stmt.setString(2, modelo.getNome());
         Calendar tmp3 = modelo.getNascimento();
         stmt.setDate(3, new Date(tmp3.getTimeInMillis()));
         stmt.setString(4, modelo.getSexo());
-        stmt.setString(5, modelo.getLogin());
-        stmt.setString(6, modelo.getSenha());
+        stmt.setString(5, modelo.getEnfermeiro());
         return stmt.executeUpdate() != 0;
 
     }
 
     @Override
-    public boolean remove(Enfermeiro modelo) throws SQLException {
+    public boolean remove(Paciente modelo) throws SQLException {
+
         if (searchBy("id", modelo.getId()).isEmpty()) {
             return false;
         }
 
-        PreparedStatement stmt = con.prepareStatement("delete from enfermeiro where id = ?");
+        PreparedStatement stmt = con.prepareStatement("delete from paciente where id = ?");
         stmt.setString(1, modelo.getId());
         return stmt.execute();
 
     }
 
     @Override
-    public boolean update(Enfermeiro modelo) throws SQLException {
+    public boolean update(Paciente modelo) throws SQLException {
+
         if (searchBy("id", modelo.getId()).isEmpty()) {
             return false;
         }
 
-        PreparedStatement stmt = con.prepareStatement("update enfermeiro set nome = ?, nascimento = ?, sexo = ?, login = ?, senha = ? where id = ?");
+        PreparedStatement stmt = con.prepareStatement("update paciente set nome = ?, nascimento = ?, sexo = ?, enfermeiro = ? where id = ?");
         stmt.setString(1, modelo.getNome());
         Calendar tmp2 = modelo.getNascimento();
         stmt.setDate(2, new Date(tmp2.getTimeInMillis()));
         stmt.setString(3, modelo.getSexo());
-        stmt.setString(4, modelo.getLogin());
-        stmt.setString(5, modelo.getSenha());
-        stmt.setString(6, modelo.getId());
+        stmt.setString(4, modelo.getEnfermeiro());
+        stmt.setString(5, modelo.getId());
         return stmt.executeUpdate() != 0;
 
     }
 
     @Override
-    public List<Enfermeiro> selectAll() throws SQLException {
-        ArrayList<Enfermeiro> lista = new ArrayList<>();
-        PreparedStatement stmt = con.prepareStatement("select * from enfermeiro");
+    public List<Paciente> selectAll() throws SQLException {
+        ArrayList<Paciente> lista = new ArrayList<>();
+        PreparedStatement stmt = con.prepareStatement("select * from paciente");
         ResultSet rs = stmt.executeQuery();
         while (rs.next()) {
             String id = rs.getString("id");
@@ -75,19 +75,18 @@ public class EnfermeiroDAO<Object> extends DAO<Enfermeiro> {
             data1.setTime(rs.getDate("nascimento"));
             Calendar nascimento = data1;
             String sexo = rs.getString("sexo");
-            String login = rs.getString("login");
-            String senha = rs.getString("senha");
-            Enfermeiro obj = new Enfermeiro(id, nome, nascimento, sexo, login, senha);
+            String enfermeiro = rs.getString("enfermeiro");
+            Paciente obj = new Paciente(id, nome, nascimento, sexo, enfermeiro);
             lista.add(obj);
         }
         return lista;
     }
 
     @Override
-    public List<Enfermeiro> searchBy(String campo, String valor) throws SQLException {
-        ArrayList<Enfermeiro> lista = new ArrayList<>();
-        PreparedStatement stmt = con.prepareStatement("select * from enfermeiro where " + campo + " = ?");
-        stmt.setString(1, valor);
+    public List<Paciente> searchBy(String tipo, String coisa) throws SQLException {
+        ArrayList<Paciente> lista = new ArrayList<>();
+        PreparedStatement stmt = con.prepareStatement("select * from paciente where " + tipo + " = ?");
+        stmt.setString(1, coisa);
         ResultSet rs = stmt.executeQuery();
         while (rs.next()) {
             String id = rs.getString("id");
@@ -96,9 +95,8 @@ public class EnfermeiroDAO<Object> extends DAO<Enfermeiro> {
             data1.setTime(rs.getDate("nascimento"));
             Calendar nascimento = data1;
             String sexo = rs.getString("sexo");
-            String login = rs.getString("login");
-            String senha = rs.getString("senha");
-            Enfermeiro obj = new Enfermeiro(id, nome, nascimento, sexo, login, senha);
+            String enfermeiro = rs.getString("enfermeiro");
+            Paciente obj = new Paciente(id, nome, nascimento, sexo, enfermeiro);
             lista.add(obj);
         }
         return lista;
