+++ Códigos/DiagnosticoSaude/src/Dao/SolucaoDAO.java
@@ -1,5 +1,6 @@
-package Model;
+package Dao;
 
+import Model.Solucao;
 import java.sql.Connection;
 import java.sql.SQLException;
 import java.sql.PreparedStatement;
